add API call and hide id in table
