+++ components/pplhp/LHPDraftPage.tsx
@@ -12,7 +12,7 @@ export default function LHPDraftPage({ linkData }) {
   // Get Laporan Hasil Pemeriksaan data
   return (
     <div className="flex flex-col md:flex-row justify-between w-full h-screen gap-36 md:gap-0">
-      <div className="md:w-3/5 h-fit px-16 space-y-6">
+      <div className="md:w-full h-fit px-16 space-y-6">
         <div className="flex flex-row text-2xl font-medium space-x-0 cursor-pointer">
           <div
             className={`flex flex-col items-end ${
