+++ components/pplhp/PPLHPDraftDataTable.tsx
@@ -60,7 +60,7 @@ const PPLHPDataTable: FC<PPLHPDataTableProps> = ({ data }) => {
   const router = useRouter();
   const [sorting, setSorting] = useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
-  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
+  const [columnVisibility, setColumnVisibility] = useState({ _id: false });
   const [rowSelection, setRowSelection] = useState({});
 
   const table = useReactTable({
